+++ src/main/java/com/nred/nuclearcraft/compat/cct/FissionPeripheral.java
@@ -14,6 +14,7 @@ import org.jspecify.annotations.Nullable;
 
 import static com.nred.nuclearcraft.helpers.Location.ncLoc;
 
+@SuppressWarnings("unused")
 public final class FissionPeripheral extends MultiblockPeripheral<FissionReactor> implements IPeripheral {
     public FissionPeripheral(FissionComputerPortEntity turbineComputerPort) {
         super(turbineComputerPort);