+++ src/main/java/com/nred/nuclearcraft/recipe/RecipeInfo.java
@@ -1,7 +1,5 @@
 package com.nred.nuclearcraft.recipe;
 
-import com.nred.nuclearcraft.handler.SizedChanceFluidIngredient;
-import com.nred.nuclearcraft.handler.SizedChanceItemIngredient;
 import it.unimi.dsi.fastutil.ints.IntList;
 
 import javax.annotation.Nonnull;