+++ src/main/java/com/nred/nuclearcraft/recipe/exchanger/HeatExchangerRecipe.java
@@ -3,7 +3,7 @@ package com.nred.nuclearcraft.recipe.exchanger;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.nred.nuclearcraft.handler.SizedChanceFluidIngredient;
+import com.nred.nuclearcraft.recipe.SizedChanceFluidIngredient;
 import com.nred.nuclearcraft.recipe.BasicRecipe;
 import com.nred.nuclearcraft.util.StreamCodecsHelper;
 import io.netty.buffer.ByteBuf;